+++ src/HoneybeeDotNet/Model/ScheduleDay.cs
@@ -43,7 +43,7 @@ namespace HoneybeeDotNet.Model
         /// <param name="type">type (default to &quot;ScheduleDay&quot;).</param>
         /// <param name="times">A list of lists with each sub-list possesing 2 values for [hour, minute]. The length of the master list must match the length of the values list. Each time in the master list represents the time of day that the corresponding value begins to take effect. For example [(0,0), (9,0), (17,0)] in combination with the values [0, 1, 0] denotes a schedule value of 0 from 0:00 to 9:00, a value of 1 from 9:00 to 17:00 and 0 from 17:00 to the end of the day. Note that this representation of times as the \&quot;time of beginning\&quot; is a different convention than EnergyPlus, which uses \&quot;time until\&quot;..</param>
         /// <param name="interpolate">Boolean to note whether values in between times should be linearly interpolated or whether successive values should take effect immediately upon the beginning time corrsponding to them. (default to false).</param>
-        public ScheduleDay(string name, List<decimal> values, string type = "ScheduleDay", List<List<int>> times = default, bool interpolate = false)
+        public ScheduleDay(string name, List<double> values, string type = "ScheduleDay", List<List<int>> times = default, bool interpolate = false)
         {
             // to ensure "name" is required (not null)
             if (name == null)
@@ -100,7 +100,7 @@ namespace HoneybeeDotNet.Model
         /// <value>A list of floats or integers for the values of the schedule. The length of this list must match the length of the times list.</value>
         [DataMember(Name="values", EmitDefaultValue=false)]
         [JsonProperty("values")]
-        public List<decimal> Values { get; set; }
+        public List<double> Values { get; set; }
 
         /// <summary>
         /// Gets or Sets Type
